requirements.txt
anthropic==0.39.0
python-dotenv==1.0.0
pymupdf==1.23.8
python-docx==1.1.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
chromadb>=0.4.0
sentence-transformers>=2.2.0
torch>=2.0.0
